update demo_frankmocap output
