+++ bodymocap/utils/imutils.py
@@ -224,36 +224,6 @@ def flip_pose(pose):
     return pose
 
 
-
-def anthro_crop_fromRaw(rawimage, bbox_XYXY):
-    bbox_w = bbox_XYXY[2] - bbox_XYXY[0]
-    bbox_h = bbox_XYXY[3] - bbox_XYXY[1]
-    bbox_size = max(bbox_w, bbox_h)     #take the max
-    bbox_center = (bbox_XYXY[:2] + bbox_XYXY[2:])*0.5
-    pt_ul = (bbox_center - bbox_size*0.5).astype(np.int32)
-    pt_br = (bbox_center + bbox_size*0.5).astype(np.int32)
-    croppedImg = rawimage[pt_ul[1]:pt_br[1], pt_ul[0]:pt_br[0]]
-    croppedImg = np.ascontiguousarray(croppedImg)
-    return rawimage, pt_ul, pt_br
-
-def anthro_convert_smpl_to_bbox(data3D, scale, trans, bbox_max_size):
-    resnet_input_size_half = bbox_max_size *0.5
-
-    data3D *= scale           #apply scaling
-    # data3D[:,0] += data3D[b,1]        #apply translation x
-    # data3D[:,1] += data3D[b,2]        #apply translation y
-    data3D[:,0:2] += trans
-    data3D*= resnet_input_size_half         #112 is originated from hrm's input size (224,24)
-
-    return data3D
-
-# def anthro_convert_bbox_to_oriIm(data3D, boxScale_o2n, bboxTopLeft, imgSizeW, imgSizeH):
-def anthro_convert_bbox_to_oriIm(pred_vert_vis, rawImg_w, rawImg_h, bbox_pt_ul, bbox_max_size):
-    pred_vert_vis[:,:2] +=  bbox_pt_ul - np.array((rawImg_w, rawImg_h))*0.5 +(bbox_max_size*0.5)  # + resnet_input_size_half#+ resnet_input_size_half 
-    return pred_vert_vis
-
-
-
 """ Extract bbox information """
 def bbox_from_openpose(openpose_file, rescale=1.2, detection_thresh=0.2):
     """Get center and scale for bounding box from openpose detections."""
