+++ renderer/od_renderer.py
@@ -33,11 +33,12 @@ class OpendrRenderer(object):
 
         input_size = 500
 
-        f = 5
+        f = 10
+
         verts[:, 0] = (verts[:, 0] - input_size) / input_size
         verts[:, 1] = (verts[:, 1] - input_size) / input_size
 
-        verts[:, 2] /= (4 * 112)
+        verts[:, 2] /= (5 * 112)
         verts[:, 2] += f
 
         cam_for_render = np.array([f, 1, 1]) * input_size
